Add conditional html attributes
Some attributes, notably ones on <source>, are conditional. Previously
they were skipped. This made using <source> impossible in <picture> as
it was lacking srcset attribute.

This commit includes additional attributes in html-sys and html. Their
conditions are not checked at type-level; this is left for a future
commit. The generated code includes the condition in the docstring.